+++ backend/src/scripts/clearData/clearBooks.js
@@ -3,12 +3,13 @@
  * @file /docman/backend/src/scripts/clearData/clearBooks.js
  * @description Script to clear all books from the database
  * @author Richard Bakos
- * @version 2.2.1
+ * @version 2.2.2
  * @license UNLICENSED
  */
 import mongoose from 'mongoose';
 import { connectDB } from '../../config/db.js';
 import Book from '../../models/Book.js';
+import { assertSafeClear } from '../lib/assertSafeClear.js';
 
 /**
  * Clear all books from the database
@@ -18,6 +19,7 @@ import Book from '../../models/Book.js';
  */
 const clearBooks = async () => {
     try {
+        assertSafeClear('clearBooks');
         await connectDB();
         
         const result = await Book.deleteMany({});
