+++ backend/src/scripts/clearData/clearNotifications.js
@@ -3,12 +3,13 @@
  * @file /docman/backend/src/scripts/clearData/clearNotifications.js
  * @description Script to clear all notifications from the database
  * @author Richard Bakos
- * @version 2.2.1
+ * @version 2.2.2
  * @license UNLICENSED
  */
 import mongoose from 'mongoose';
 import { connectDB } from '../../config/db.js';
 import Notification from '../../models/Notification.js';
+import { assertSafeClear } from '../lib/assertSafeClear.js';
 
 /**
  * Clear all notifications from the database
@@ -18,6 +19,7 @@ import Notification from '../../models/Notification.js';
  */
 const clearNotifications = async () => {
     try {
+        assertSafeClear('clearNotifications');
         await connectDB();
         
         const result = await Notification.deleteMany({});
